New attempt at a better BM3D algo
This new attempt follows the work of the repository cited in the first markdown of the notebook. It improves the other one by introducing new transformations (such as the hadamard) and significantly reduced the computational time by operating directly on matrices instead of using nested for loops.
Still have to code the second step of the algorithm, but so far the results are promising.