Upload final implementation of BM3D
This new implementation has been way more taxing on my personal computer, probably because of the increased quantity of memory to keep close to the processor in order to operate on the input image. Hence it has been run succesfully on colab, even though it as been writteon on my VScode.
This version works way better than the other one both in terms of PSNR of the result and quickness on producing the result.